requirements.txt
# Document Parsing & Chunking
pypdf
langchain-text-splitters

# Vector Database & Embeddings
qdrant-client
langchain-qdrant
fastembed

# AI Agents & LLM (Gemini)
langchain-core
langchain-community
langchain-google-genai
langgraph

# Financial Data & Web Search
yfinance
duckduckgo-search

# Environment, Data Validation & UI
python-dotenv
pydantic
streamlit
plotly